+++ RevenuePlanner/Helpers/Common.cs
@@ -5688,7 +5688,7 @@ namespace RevenuePlanner.Helpers
 
                 if (goalType != "" && goalValue != "" && goalValue != "0")
                 {
-                    double inputValue = Convert.ToInt64(goalValue.Trim().Replace(",", "").Replace(Sessions.PlanCurrencySymbol, "")); //Modified by Rahul Shah for PL #2499
+                    double inputValue = Convert.ToDouble(goalValue.Trim().Replace(",", "").Replace(Sessions.PlanCurrencySymbol, "")); //Modified by Rahul Shah for PL #2499
 
                     if (goalType == Enums.PlanGoalType.INQ.ToString())
                     {
